% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmemes.R
\name{show_me_an_R_meme}
\alias{show_me_an_R_meme}
\title{Display an R Meme at your convenience}
\usage{
show_me_an_R_meme(meme_index = NULL, random_meme = T)
}
\arguments{
\item{meme_index}{Index number of the meme that you want to be displayed. Defaults to \code{NULL}.}

\item{random_meme}{Show a random meme. Defaults to \code{TRUE}.}
}
\description{
This function shows you an R Meme from the R Memes for Statistical Fiends Facebook page
}
\examples{

## show a random R meme
show_me_an_R_meme()

## show a specific R meme
show_me_an_R_meme(179)

}
